small review fixes during test
