+++ .pylintrc
@@ -1,3 +1,6 @@
+[MASTER]
+init-hook='import sys; import os; sys.path.append(os.path.dirname(os.path.abspath(__file__)))'
+
 [DESIGN]
 min-public-methods=1
 max-attributes=10
