+++ pyproject.toml
@@ -2,7 +2,7 @@
 name = "my_chat_gpt"
 version = "0.1.0"
 description = "A collection of LLM documents and experiments, including ChatGPT and local Ollama."
-authors = ["Pieter Kuppens <pieter.kuppens@gmail.com>"]
+authors = [{name = "Pieter Kuppens <pieter.kuppens@gmail.com>"}]
 license = {name = "MIT", file = "LICENSE"}
 
 dependencies = [
@@ -228,7 +228,7 @@ dependencies = [
     "SQLAlchemy==2.0.36",
     "sqlite-anyio==0.2.3",
     "stack-data==0.6.3",
-    "starlette==0.45.2",
+    "starlette==0.37.2",
     "sympy==1.13.1",
     "tenacity==9.0.0",
     "terminado==0.18.1",
@@ -266,12 +266,12 @@ dependencies = [
     "yarl==1.18.3",
     "ypy-websocket==0.12.4",
     "zipp==3.21.0",
-    "local_utils"
 ]
 
 [build-system]
-requires = ["uv"]
-build-backend = "uv.build"
+requires = ["setuptools>=42", "wheel", "uv"]
+build-backend = "setuptools.build_meta"
+# build-backend = "uv.build"
 
 [tool.black]
 line-length = 132
