+++ setup.py
@@ -1,0 +1,7 @@
+from setuptools import setup, find_packages
+
+setup(
+    name="my_chat_gpt",
+    version="0.1.0",
+    packages=find_packages(),
+)
